+++ omq-tokio/src/socket/actor/mod.rs
@@ -200,7 +200,7 @@ pub(crate) struct SocketDriver {
     closing: bool,
     close_deadline: Option<Instant>,
     close_ack: Option<oneshot::Sender<Result<()>>>,
-    spsc: super::handle::SpscHandles,
+    spsc: super::recv::SpscHandles,
     wire_slot: super::handle::WireSlotHolder,
     rr_slots: super::handle::RrSlots,
     compression_pool: Option<Arc<crate::engine::compression_pool::CompressionPool>>,
@@ -217,7 +217,7 @@ impl SocketDriver {
         cancel: CancellationToken,
         monitor: MonitorPublisher,
         send_strategy: SendStrategy,
-        spsc: super::handle::SpscHandles,
+        spsc: super::recv::SpscHandles,
         type_state: Arc<Mutex<TypeState>>,
         req_awaiting_reply: Arc<AtomicBool>,
         wire_slot: super::handle::WireSlotHolder,
