+++ omq-tokio/src/socket/mod.rs
@@ -10,6 +10,7 @@ pub(crate) mod actor;
 pub(crate) mod dispatch;
 pub mod handle;
 pub mod monitor;
+pub(crate) mod recv;
 pub(crate) mod type_state;
 pub(crate) mod udp;
